requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
idna==3.9
importlib_metadata==8.5.0
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
packaging==24.1
platformdirs==4.3.3
pluggy==1.5.0
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
tomli==2.0.1
urllib3==2.2.3
Werkzeug==3.0.4
xmltodict==0.13.0
zeep==4.2.1
zipp==3.20.2
